### 2023-09-16 #5851 @dfm
@musoke — Thanks for your submission. I'm the managing editor of this track and I'm working to assign a suitable editor for this review. Everyone is working at capacity at the moment so it might take a few weeks before we can start the review, but I'll keep you posted.

In the meantime:

- Please make sure you add a "Statement of need" section to the paper, and fix the syntax error in the paper metadata. I think that the issue might be the `&` in your affiliation, which is a special character in YAML. Perhaps you can fix it by quoting your affiliation string. Check out the [JOSS GitHub Actions workflow](https://github.com/marketplace/actions/open-journals-pdf-generator) for testing.
- You might also consider expanding the documentation while you wait. From a quick look, I'd say that the example notebooks should have more textual explanation included to make them more useful. As it stands, the docs are somewhat borderline with respect to the JOSS requirements.

Please don't hesitate to let me know if you have any questions as the review progresses!
