### 2023-09-16 2277764
doc(joss): Fix syntax in metadata
https://github.com/openjournals/joss-reviews/issues/5851#issuecomment-1722257091
