### 2023-09-16 83fb99e
doc(joss): Fix metadata in paper
