### 2023-09-16 10a19f0
doc(joss): Fix date
