+++ joss/paper.md
@@ -14,7 +14,7 @@ authors:
 affiliations:
  - name: "Department of Physics & Astronomy, University of New Hampshire, USA"
    index: 1
-date: 1 June 2023
+date: 15 September 2023
 bibliography: paper.bib
 
 ---
