+++ joss/paper.md
@@ -10,7 +10,7 @@ tags:
 authors:
   - name: Nathan Musoke
     orcid:  0000-0001-9839-9256
-    affiliation: "1" # (Multiple affiliations must be quoted)
+    affiliation: 1
 affiliations:
  - name: "Department of Physics and Astronomy, University of New Hampshire, USA"
    index: 1
