+++ Project.toml
@@ -15,6 +15,7 @@ NPZ = "15e1cf62-19b3-5cfa-8e77-841668bca605"
 PencilArrays = "0e08944d-e94e-41b1-9406-dcf66b6a9d2e"
 PencilFFTs = "4a48f351-57a6-4416-9ec4-c37015456aae"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+Strided = "5e0ebb24-38b0-5f93-81fe-25c709ecae67"
 
 [compat]
 AbstractFFTs = "0.5, 1.0"
@@ -28,4 +29,5 @@ NPZ = "0.4"
 PencilArrays = "0.11, 0.12, 0.13, 0.14, 0.15, 0.16, 0.17, 0.18"
 PencilFFTs = "0.12, 0.13, 0.14, 0.15"
 Statistics = "1.8, 1.9"
+Strided = "2.0.4"
 julia = "1.8, 1.9"
