### 2023-09-23 d557bdd
opt: Parallelise calculation of coordinate arrays
There are functions which calculate radial and angular coordinates from
Cartesian.  Use Strided.jl to parallelise them.
