### 2023-10-09 #5851 @xuanxu
@musoke The editorial bot can't compile the paper because it's getting confused it with another existing git reference: it finds the `paper.md` branch when looking for the `paper.md` file. Can you rename that branch? any other name will do. `joss`, `joss-paper` or `joss-review` are common alternatives.
