### 2023-10-20 833e08a
chore: Add compat entries for std libs
The general Julia registry will soon require that all standard library
dependencies have compat entries.  This will allow them to be
upgradeable from Julia 1.10 onwards.

https://discourse.julialang.org/t/psa-compat-requirements-in-the-general-registry-are-changing/104958

Add them.
