+++ Project.toml
@@ -18,11 +18,14 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
 AbstractFFTs = "0.5, 1.0"
+Dates = "1.8, 1.9"
 FFTW = "1.2"
 Folds = "0.2.8"
 HDF5 = "0.16, 0.17"
+LinearAlgebra = "1.8, 1.9"
 MPI = "0.20"
 NPZ = "0.4"
 PencilArrays = "0.11, 0.12, 0.13, 0.14, 0.15, 0.16, 0.17, 0.18"
 PencilFFTs = "0.12, 0.13, 0.14, 0.15"
+Statistics = "1.8, 1.9"
 julia = "1.8, 1.9"
