### 2023-11-08 #6035 @editorialbot
[REVIEW]: UltraDark.jl: A Julia package for simulation of cosmological scalar fields
**Submitting author:** @musoke (<a href="http://orcid.org/0000-0001-9839-9256">Nathan Musoke</a>)
**Repository:** https://github.com/musoke/UltraDark.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.9.7
