### 2023-11-09 365cfc0
fixup: Add missing doi to reference
