+++ joss/paper.bib
@@ -198,7 +198,6 @@
   doi           = {10.1103/PhysRevLett.124.061301},
   eprint        = {1909.11678},
   primaryclass  = {astro-ph.CO},
-  readstatus    = {read},
 }
 
 @inproceedings{padmanabhan2020simulating,
@@ -207,7 +206,8 @@
   booktitle={2020 IEEE International Parallel and Distributed Processing Symposium Workshops (IPDPSW)},
   pages={678--678},
   year={2020},
-  organization={IEEE}
+  organization={IEEE},
+  doi= {10.1109/ipdpsw50202.2020.00120},
 }
 
 @misc{folds,
