### 2023-11-09 0c88441
fixup: Add JCAP details to vortices paper
