+++ joss/paper.bib
@@ -58,16 +58,21 @@
 
 @Article{Glennon:2023oqa,
   author        = {Glennon, Noah and Mirasola, Anthony E. and Musoke, Nathan and Neyrinck, Mark C. and Prescod-Weinstein, Chanda},
+  journal       = {Journal of Cosmology and Astroparticle Physics},
   title         = {Scalar dark matter vortex stabilization with black holes},
   year          = {2023},
-  month         = {1},
+  month         = {jul},
+  number        = {07},
+  pages         = {004},
+  volume        = {2023},
   archiveprefix = {arXiv},
-  copyright     = {Creative Commons Attribution Non Commercial No Derivatives 4.0 International},
-  doi           = {10.48550/ARXIV.2301.13220},
+  copyright     = {arXiv.org perpetual, non-exclusive license},
+  date          = {2023-01-30},
+  doi           = {10.1088/1475-7516/2023/07/004},
   eprint        = {2301.13220},
   keywords      = {Cosmology and Nongalactic Astrophysics (astro-ph.CO), Astrophysics of Galaxies (astro-ph.GA), High Energy Physics - Phenomenology (hep-ph), FOS: Physical sciences},
   primaryclass  = {astro-ph.CO},
-  publisher     = {arXiv},
+  publisher     = {{IOP} Publishing},
 }
 
 @article{Julia-2017,
