### 2023-11-22 #6035 @kiranshila
# Review

Great work @musoke, nice to see more Julia work in astronomy! The fact that you have published science results using your package is really awesome!

I have just a few comments,

## Critical
1. Please add something like a `CONTRIBUTING` file to the repository to satisfy the "Community Guidelines" documentation requirement.

## General Comments / Feedback
1. I think the API docs are a little on the light side, most of the "how do I use this package" stuff seems to be contained in the notebooks. Maybe another page or so on general use would be nice. Still seems like more than `PyUltraLight` though :) I'm not in this domain, so it's hard for me to say if some of the use is obvious, but I thought I'd mention it anyway.
2. It might be nice to have a result figure or so for eye candy
