### 2023-12-02 #6035 @szabo137
# Review

First of all, very nice work @musoke, I also love to see more of Julia in astronomy!

Regarding the package, I also want to add some comments:

### Critical
* IMHO the package's documentation deserves a bit more content to highlight its functionality and context. Maybe you could consider adding a bit more explanation on what your package does, and maybe a little quickstart section would help to get in touch with the code
* I also suggest adding some guide (either as a file or in the docs, or both) on how to interact with the repository, if one wants to contribute, i.e. following the community guideline requirements.

### General feedback
I like the simplistic API design and tight interconnection of this package with the scientific use case.  Especially, the addition of the example folder helps a lot to understand the package and the science behind it.
