### 2023-12-05 2493dcf
ci: Update julia-format action to v2
