### 2024-01-12 musoke/UltraDark.jl#167 @musoke
Contribution guidelines
From https://github.com/openjournals/joss-reviews/issues/6035#issuecomment-1837265250:

> I also suggest adding some guide (either as a file or in the docs, or both) on how to interact with the repository, if one wants to contribute, i.e. following the community guideline requirements.

and https://github.com/openjournals/joss-reviews/issues/6035#issuecomment-1823414644

> Please add something like a CONTRIBUTING file to the repository to satisfy the "Community Guidelines" documentation requirement.
