### 2024-01-12 musoke/UltraDark.jl#168 @musoke
Expand documentation
https://github.com/openjournals/joss-reviews/issues/6035#issuecomment-1837265250:

> IMHO the package's documentation deserves a bit more content to highlight its functionality and context. Maybe you could consider adding a bit more explanation on what your package does, and maybe a little quickstart section would help to get in touch with the code

https://github.com/openjournals/joss-reviews/issues/6035#issuecomment-1823414644:

> I think the API docs are a little on the light side, most of the "how do I use this package" stuff seems to be contained in the notebooks. Maybe another page or so on general use would be nice. Still seems like more than PyUltraLight though :) I'm not in this domain, so it's hard for me to say if some of the use is obvious, but I thought I'd mention it anyway.
