### 2024-01-12 musoke/UltraDark.jl#169 @musoke
Add a result figure to JOSS draft
https://github.com/openjournals/joss-reviews/issues/6035#issuecomment-1823414644

> 2. It might be nice to have a result figure or so for eye candy
