+++ CONTRIBUTING.md
@@ -1,0 +1,16 @@
+# Contribution guidelines
+
+## Contributing code
+
+Open a pull request in this repository if you have code to contribute.
+For efficiency, you may want to open an issue and seek feedback before submitting non-trivial changes.
+
+### Code style
+
+This repository uses [JuliaFormatter.jl](https://github.com/domluna/JuliaFormatter.jl) to enforce consistent formatting.
+
+
+## Reporting issues or problems
+
+Please open an issue in this repository if you run into a problem running UltraDark or want to request a feature.
+Include a [minimal reproducible example](https://en.wikipedia.org/wiki/Minimal_reproducible_example) with any bug reports.
