### 2024-01-15 d42cead
doc: CONTRIBUTING
Add a minimal guide for contributions and issues.

Fixes: https://github.com/musoke/UltraDark.jl/issues/167
