### 2024-01-15 823c1de
doc: Add code of conduct
From https://www.contributor-covenant.org/
