### 2024-01-15 2f59526
ci: Bump Julia to 1.10
