+++ .github/workflows/ci.yml
@@ -17,6 +17,7 @@ jobs:
         version:
           - '1.8'
           - '1.9'
+          - '1.10'
         os:
           - ubuntu-latest
           - macOS-latest
