### 2024-01-16 58ec664
ci: actions/cache -> julia-actions/cache
Use the new Julia-specific cache action instead of the one included from
PkgTemplates.jl.
