### 2024-01-26 786dbc6
opt: Use PrecomileTools
Use PrecompileTools to precompile functions required for a very basic
simulation.  This should help with TTFX.

Could be extended to `PencilGrids`, functions inside `simulate!` that
only run intermittently, and summary statistics.
