+++ test/sims/scale_factor.jl
@@ -8,8 +8,6 @@ using PencilFFTs
 
 resol = 64
 
-include("../../examples/init_soliton.jl")
-
 output_dir = mktempdir()
 
 for a in [UltraDark.constant_scale_factor, t -> t]
@@ -29,13 +27,13 @@ for a in [UltraDark.constant_scale_factor, t -> t]
 
     grids = Grids(10.0, resol)
 
-    mass = 10
-    position = [0, 0, 0]
-    velocity = [0, 0, 0]
-    phase = 0
-    t0 = 0
+    mass = 10.0
+    position = [0.0, 0.0, 0.0]
+    velocity = [0.0, 0.0, 0.0]
+    phase = 0.0
+    t0 = 0.0
 
-    add_soliton(grids, mass, position, velocity, phase, t0)
+    UltraDark.Initialise.add_fdm_soliton!(grids, mass, position, velocity, phase, t0)
 
     @test simulate!(grids, options, output_config) == nothing
 
