### 2024-01-26 1060075
feat: Soliton initialisation
Previously soliton initialisation was in a script which was imported in
some examples.  Move this into the package as a function so it is
properly documented and tested.

Add this function to precompilation.
