### 2024-01-26 7d53588
chore: Bump version to 0.9.6
