### 2024-01-26 38b6c3c
chore: Add Julia v1.10 to compat
This should have been included with the update to CI targets in
2f5952666798553f9d0dc5bc90f9e90120c9c290.
