+++ Project.toml
@@ -32,4 +32,4 @@ PencilFFTs = "0.12, 0.13, 0.14, 0.15"
 PrecompileTools = "1.2"
 Statistics = "1.8, 1.9"
 Strided = "2.0.4"
-julia = "1.8, 1.9"
+julia = "1.8, 1.9, 1.10"
