### 2024-01-26 03bba0b
doc: Update email address, dates
