+++ Project.toml
@@ -1,6 +1,6 @@
 name = "UltraDark"
 uuid = "1c8d022d-dfc0-4b41-80ab-3fc7e88cdfea"
-authors = ["Nathan Musoke <n.musoke@auckland.ac.nz>"]
+authors = ["Nathan Musoke <nathan.musoke@gmail.com>"]
 version = "0.9.6"
 
 [deps]
