### 2024-01-26 b52d7f9
doc: Add page for initialisation
