+++ docs/src/man/init.md
@@ -1,0 +1,5 @@
+# Soliton initialisation
+
+```@autodocs
+Modules = [UltraDark.Initialise]
+```
