### 2024-01-26 c9f8ecb
doc: Add page for initialisation
