+++ .github/workflows/ci.yml
@@ -47,7 +47,9 @@ jobs:
           # Fix for Plots.jl with GR backend.
           GKSwstype: nul
       - uses: julia-actions/julia-processcoverage@v1
-      - uses: codecov/codecov-action@v3
+      - uses: codecov/codecov-action@v4
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         with:
           file: lcov.info
       # - uses: julia-actions/julia-uploadcoveralls@v1
