### 2024-03-06 a3433c0
doc(joss): Add ref to dynamical friction PRD2023
