### 2024-03-06 62c0623
joss: Add reference to PRD 2023
