+++ joss/paper.bib
@@ -11,11 +11,18 @@
   primaryclass  = {astro-ph.CO},
 }
 
-@Article{2023xxx,
-  author        = {Glennon, Noah and Musoke, Nathan and Nadler, Ethan O. and Prescod-Weinstein, Chanda and Wechsler, Risa H.},
-  title         = {Dynamical friction in self-interacting ultralight dark matter},
-  journal       = {under review},
-  year          = {2023},
+@article{Glennon:2023gfm,
+    author = "Glennon, Noah and Musoke, Nathan and Nadler, Ethan O. and Prescod-Weinstein, Chanda and Wechsler, Risa H.",
+    title = "{Dynamical friction in self-interacting ultralight dark matter}",
+    eprint = "2312.07684",
+    archivePrefix = "arXiv",
+    primaryClass = "astro-ph.CO",
+    doi = "10.1103/PhysRevD.109.063501",
+    journal = "Phys. Rev. D",
+    volume = "109",
+    number = "6",
+    pages = "063501",
+    year = "2024"
 }
 
 @Article{Glennon:2022huu,
