### 2024-03-06 a600bbd
doc(joss): Add ref to dynamical friction PRD2023
