### 2024-03-26 ae67a16
doc: add DocumenterInterLinks
Add DocumenterInterLinks, which allows linking to the documentation of
other projects.  This requires updating to Documenter 1.3.0.
