+++ docs/Project.toml
@@ -1,6 +1,8 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+DocumenterInterLinks = "d12716ef-a0f6-4df4-a9f1-a5a34e75c656"
 UltraDark = "1c8d022d-dfc0-4b41-80ab-3fc7e88cdfea"
 
 [compat]
-Documenter = "0.24"
+Documenter = "1"
+DocumenterInterLinks = "0.3"
