### 2024-03-26 f40f795
doc: add DocumenterInterLinks
Add DocumenterInterLinks, which allows linking to the documentation of
other projects.  This requires updating to Documenter 1.3.0.
