### 2024-03-27 3d70f37
doc: Add overview of a simple simulation
Fixes: https://github.com/musoke/UltraDark.jl/issues/168
