+++ docs/Project.toml
@@ -1,6 +1,9 @@
 [deps]
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterInterLinks = "d12716ef-a0f6-4df4-a9f1-a5a34e75c656"
+LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
+NPZ = "15e1cf62-19b3-5cfa-8e77-841668bca605"
 UltraDark = "1c8d022d-dfc0-4b41-80ab-3fc7e88cdfea"
 
 [compat]
