### 2024-03-27 e770ad3
doc: Add overview of a simple simulation
Fixes: https://github.com/musoke/UltraDark.jl/issues/168
