### 2024-04-01 b9999de
doc: Consolidate documentation for Grids
`Grids` and `PencilGrids` have separate docstrings for different
constructor methods.  Consolidate these.
