### 2024-04-01 55919a2
doc: Consolidate documentation for Grids
`Grids` and `PencilGrids` have separate docstrings for different
constructor methods.  Consolidate these.
