### 2024-04-12 02387d1
doc: Expand intro describing what UltraDark does
Fixes: https://github.com/musoke/UltraDark.jl/issues/175
Related: https://github.com/musoke/UltraDark.jl/issues/168
