### 2024-04-12 8018d2d
doc(joss): Add an example figure to JOSS paper
Fixes: https://github.com/musoke/UltraDark.jl/issues/169
