### 2024-04-12 d28d497
doc(joss): Add an example figure to JOSS paper
Fixes: https://github.com/musoke/UltraDark.jl/issues/169
