+++ versions.js
@@ -11,3 +11,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
+var DOCUMENTER_NEWEST = "v0.9.6";
+var DOCUMENTER_STABLE = "stable";
