+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-1","page":"API","title":"API","text":"","category":"section"},{"location":"api/#Index-1","page":"API","title":"Index","text":"","category":"section"},{"location":"api/#","page":"API","title":"API","text":"","category":"page"},{"location":"api/#Docstrings-1","page":"API","title":"Docstrings","text":"","category":"section"},{"location":"api/#","page":"API","title":"API","text":"Modules = [UltraDark, UltraDark.Config, UltraDark.Summary]","category":"page"},{"location":"api/#UltraDark.Grids","page":"API","title":"UltraDark.Grids","text":"struct containing grids used in a simulation\n\nExamples\n\njulia> using UltraDark\n\njulia> len = 1;\n\njulia> resol = 16;\n\njulia> Grids(len, resol);\n\n\n\n\n\n\n","category":"type"},{"location":"api/#UltraDark.Grids-Tuple{Any, Int64}","page":"API","title":"UltraDark.Grids","text":"Grids(length, resol::Int)\n\nConstructor for Grids\n\nCreate an empty grid with length length and resolution resol\n\nExamples\n\njulia> using UltraDark\n\njulia> Grids(1.0, 64);\n\n\n\n\n\n\n","category":"method"},{"location":"api/#UltraDark.Grids-Tuple{Any, Tuple{Int64, Int64, Int64}}","page":"API","title":"UltraDark.Grids","text":"Grids(length_tuple, resol_tuple::Tuple{Int, Int, Int})\n\nConstructor for Grids\n\nCreate an empty length[1]xlength[2]xlength[3] grid with resolution resol[1]xresol[2]xresol[3].\n\nExamples\n\njulia> using UltraDark\n\njulia> Grids((1.0, 1.0, 0.5), (64, 64, 32));\n\n\n\n\n\n\n","category":"method"},{"location":"api/#UltraDark.PencilGrids","page":"API","title":"UltraDark.PencilGrids","text":"struct containing grids used in a simulation\n\nExamples\n\njulia> using UltraDark\n\njulia> len = 1;\n\njulia> resol = 16;\n\njulia> PencilGrids(len, resol);\n\n\n\n\n\n\n","category":"type"},{"location":"api/#UltraDark.PencilGrids-Tuple{Any, Int64}","page":"API","title":"UltraDark.PencilGrids","text":"PencilGrids(length, resol)\n\nConstructor for PencilGrids\n\nCreate an empty grid with length length and resolution resol.  Uses PencilFFTs to create PencilArrays.\n\nExamples\n\njulia> using UltraDark\n\njulia> PencilGrids(1.0, 64);\n\n\n\n\n\n\n","category":"method"},{"location":"api/#UltraDark.PencilGrids-Tuple{Any, Tuple{Int64, Int64, Int64}}","page":"API","title":"UltraDark.PencilGrids","text":"PencilGrids(length_tuple, resol_tuple::Tuple{Int, Int, Int})\n\nConstructor for PencilGrids\n\nCreate an empty length[1]xlength[2]xlength[3] grid with resolution resol[1]xresol[2]xresol[3]. Each grid is a PencilArray, allowing multiprocess FFTs.\n\nExamples\n\njulia> using UltraDark\n\njulia> PencilGrids((1.0, 1.0, 0.5), (64, 64, 32));\n\n\n\n\n\n\n","category":"method"},{"location":"api/#UltraDark.dV-Tuple{Any}","page":"API","title":"UltraDark.dV","text":"dV(grids)\n\nCalculate the volume of each grid cell\n\nExamples\n\njulia> usin
[…]
