### 2024-04-12 #6035 @musoke
@xuanxu, I have made the changes suggested by @kiranshila and @szabo137.  They are all contained in the latest commit on the `joss` branch, in the commit tagged [`joss-resubmission`](https://github.com/musoke/UltraDark.jl/releases/tag/joss-resubmission).  The updated documentation pages are rendered at https://musoke.github.io/UltraDark.jl/dev/.
