### 2024-04-12 3027219
doc(joss): Add missing DOIs to paper.bib
