+++ joss/paper.bib
@@ -163,21 +163,25 @@
 }
 
 @Article{Eschle:2023ikn,
-  author        = {Eschle, J. and others},
-  title         = {Potential of the {Julia} programming language for high energy physics computing},
+  author        = {Eschle, J. and Gal, T. and Giordano, M. and Gras, P. and Hegner, B. and Heinrich, L. and Acosta, U. Hernandez and Kluth, S. and Ling, J. and Mato, P. and Mikhasenko, M. and Briceño, A. Moreno and Pivarski, J. and Samaras-Tsakiris, K. and Schulz, O. and Stewart, G. . A. and Strube, J. and Vassilev, V.},
+  journal       = {Computing. Comput Softw Big Sci 7, 10 (2023)},
+  title         = {Potential of the Julia programming language for high energy physics computing},
   year          = {2023},
-  month         = {6},
+  month         = {oct},
+  number        = {1},
+  volume        = {7},
   archiveprefix = {arXiv},
+  copyright     = {Creative Commons Attribution 4.0 International},
+  date          = {2023-06-06},
+  doi           = {10.1007/s41781-023-00104-x},
   eprint        = {2306.03675},
+  keywords      = {High Energy Physics - Phenomenology (hep-ph), Programming Languages (cs.PL), High Energy Physics - Experiment (hep-ex), Computational Physics (physics.comp-ph), FOS: Physical sciences, FOS: Computer and information sciences, J.2},
   primaryclass  = {hep-ph},
+  publisher     = {Springer Science and Business Media {LLC}},
 }
 
 @misc{anim:vortex,
-  author       = {Glennon, Noah and
-                  Mirasola, Anthony E. and
-                  Musoke, Nathan and
-                  Neyrinck, Mark C. and
-                  Prescod-Weinstein, Chanda},
+  author       = {Glennon, Noah and Mirasola, Anthony E. and Musoke, Nathan and Neyrinck, Mark C. and Prescod-Weinstein, Chanda},
   title        = {{Supplementary animations for "Scalar dark matter vortex stabilization with black holes"}},
   month        = feb,
   year         = 2023,
@@ -187,9 +191,7 @@
 }
 
 @misc{anim:multi,
-  author       = {Glennon, Noah and
-                  Musoke, Nathan and
-                  Prescod-Weinstein, Chanda},
+  author       = {Glennon, Noah and Musoke, Nathan and Prescod-Weinstein, Chanda},
   title        = {{Supplementary animations for "Simulations of multi-field ultralight axion-like dark matter"}},
   month        = feb,
   year         = 2023,
@@ -199,13 +201,8 @@
 }
 
 @misc{anim:friction,
-  author       = {Glennon, Noah and
-                  Musoke, Nathan and
-                  Nadler, Ethan O. and
-                  Prescod-Weinstein, Chanda and
-                  Wechsler, Risa H.},
-  title        = {{Supplementary animations for "Dynamical friction 
-                   in self-interacting ultralight dark matter"}},
+  author       = {Glennon, Noah and Musoke, Nathan and Nadler, Ethan O. and Prescod-Weinstein, Chanda and Wechsler, Risa H.},
+  title        = {{Supplementary animations for "Dynamical friction in self-interacting ultralight dark matter"}},
   month        = dec,
   year         = 2023,
   publisher    = {Zenodo},
@@ -415,12 +412,22 @@
 
 @Article{Jain:2023qty,
   author        = {Jain, Mudit and Amin, Mustafa A. and Pu, Han},
-  title         = {i-SPin 2: An integrator for general spin-s Gross-Pitaevskii systems},
+  journal       = {Phys. Rev. E 108, 055305, 15 November 2023},
[…]
