### 2024-04-12 9d21ba6
docs: Link more objects in overview.md
