### 2024-04-13 aa9a5ee
doc: Clarify many docstrings
Clarify many of the docstrings in the package.

Document of the fields of the structs defined.

Add docstrings to the modules defined.  These are nice are section
headers in the rendered docs.

Fixes: https://github.com/musoke/UltraDark.jl/issues/185

Related: https://github.com/musoke/UltraDark.jl/issues/168
