+++ src/UltraDark.jl
@@ -2,6 +2,7 @@ module UltraDark
 
 using Base.Threads: @threads
 using AbstractFFTs: fftfreq, rfftfreq
+using DocStringExtensions
 import Folds
 using LinearAlgebra
 using PencilFFTs, MPI
@@ -190,6 +191,8 @@ function take_steps!(grids, t_start, Δt, n, output_config, a, constants, extern
 end
 
 """
+$(TYPEDSIGNATURES)
+
 Evolve `grids` forward from `t_start` to `t_end`
 """
 function evolve_to!(
@@ -234,6 +237,26 @@ function evolve_to!(
     t
 end
 
+"""
+    simulate!(grids, output_config::OutputConfig; constants = nothing, external_states = (),)
+    simulate!(grids, sim_config, output_config::OutputConfig; constants = nothing, external_states = (),)
+
+`simulate!` is the main entry point into a simulation
+
+`simulate!` evolves `grids` and `external_states` forward, writing output as specified by `output_config`.
+
+# Arguments
+
+`grids::AbstractGrids` Contains coordinate systems and the scalar field `ψ`.
+
+`sim_config::UltraDark.Config` gives further control over the way the simulation is done.
+
+`output_config::OutputConfig` controls what output is written and when.
+
+`constants::Any` can be used to overload other functions in the evolution, for example to introduce self-interactions.
+
+`external_states::Tuple{Any}` can be used in combination with overloading to add other dynamical objects to a simulation.
+"""
 function simulate!(
     grids,
     output_config::OutputConfig;
