### 2024-04-13 3965b3a
doc: Move Config and Output to own pages
The docstrings for `UltraDark.Config` and `UltraDark.Output` are
commingled and unclear.  Move them to their own pages for clarity.

Related: https://github.com/musoke/UltraDark.jl/issues/168
