+++ docs/src/man/config.md
@@ -1,0 +1,5 @@
+# Simulation Configuration
+
+```@autodocs
+Modules = [UltraDark.Config]
+```
