+++ docs/src/man/output.md
@@ -1,0 +1,5 @@
+# Output Configuration
+
+```@autodocs
+Modules = [UltraDark.Output]
+```
