+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API","page":"API","title":"API","text":"","category":"section"},{"location":"api/#Index","page":"API","title":"Index","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"","category":"page"},{"location":"api/#Docstrings","page":"API","title":"Docstrings","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"Modules = [UltraDark, UltraDark.Config, UltraDark.Output]","category":"page"},{"location":"api/#UltraDark.AbstractGrids","page":"API","title":"UltraDark.AbstractGrids","text":"AbstractGrids\n\nAbstract type for grids containing simulation data\n\n\n\n\n\n","category":"type"},{"location":"api/#UltraDark.Grids","page":"API","title":"UltraDark.Grids","text":"Grids(length, resol::Int)\nGrids(length_tuple, resol_tuple::Tuple{Int, Int, Int})\n\nstruct containing grids used in a simulation\n\nExamples\n\nCreate an empty grid with length length and resolution resol\n\njulia> using UltraDark\n\njulia> length = 1;\n\njulia> resol = 16;\n\njulia> Grids(length, resol);\n\njulia> size(ans.ψx)\n(16, 16, 16)\n\nCreate an empty length[1]xlength[2]xlength[3] grid with resolution resol[1]xresol[2]xresol[3].\n\njulia> using UltraDark\n\njulia> Grids((1.0, 1.0, 0.5), (64, 64, 32));\n\njulia> size(ans.ψx)\n(64, 64, 32)\n\n\n\n\n\n","category":"type"},{"location":"api/#UltraDark.PencilGrids","page":"API","title":"UltraDark.PencilGrids","text":"PencilGrids(length, resol)\nPencilGrids(length_tuple, resol_tuple::Tuple{Int, Int, Int})\n\nstruct containing grids used in a simulation\n\nEach grid is a PencilArray, allowing multiprocess FFTs. This comes with significant overhead so is only useful when running in a multi-node environment.\n\nExamples\n\nCreate an empty grid with length length and resolution resol.  Uses PencilFFTs to create PencilArrays.\n\njulia> using UltraDark\n\njulia> len = 1;\n\njulia> resol = 16;\n\njulia> PencilGrids(len, resol);\n\n\nCreate an empty length[1]xlength[2]xlength[3] grid with resolution resol[1]xresol[2]xresol[3].\n\njulia> using UltraDark\n\njulia> PencilGrids((1.0, 1.0, 0.5), (64, 64, 32));\n\n\n\n\n\n\n","category":"type"},{"location":"api/#UltraDark.E_grav-Tuple{Any, Any}","page":"API","title":"UltraDark.E_grav","text":"E_grav(grids)\nE_grav(grids, psi)\n\nGravitational potential energy\n\n\n\n\n\n","category":"method"},{"location":"api/#UltraDark.E_gravity_density-Tuple{Number, Real}","page":"API","title":"UltraDark.E_gravity_density","text":"E_gravity_density(psi, Phi)\n\nGravitational energy density of field psi in gravitational potential Phi\n\n\n\n\n\n","category":"method"},{"location":"api/#UltraDark.E_kq-Tuple{Any}","page":"API","title":"UltraDark.E_kq","text":"E_kq(grids)\nE_kq(grids, psi)\n\nSum of kinetic and quantum energies\n\n\n\n\n\n","category":"method"},{"location":"api/#UltraDark.E_total-Tuple{Any}","page":"API","title":"UltraDark.E_total","text":"E_total(grids; constants=nothing)\n\nTotal energy of the sc
[…]
