### 2024-04-15 #6035 @szabo137
@musoke Thanks for the update on your package and paper. It looks good to me, all issues are addressed with the update. Therefore, I recommend publishing this submission in JOSS.
