+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API","page":"API","title":"API","text":"","category":"section"},{"location":"api/#Index","page":"API","title":"Index","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"","category":"page"},{"location":"api/#Docstrings","page":"API","title":"Docstrings","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"Modules = [UltraDark]","category":"page"},{"location":"api/#UltraDark.AbstractGrids","page":"API","title":"UltraDark.AbstractGrids","text":"AbstractGrids\n\nabstract type AbstractGrids\n\nAbstract type for grids containing simulation data\n\n\n\n\n\n","category":"type"},{"location":"api/#UltraDark.Grids","page":"API","title":"UltraDark.Grids","text":"struct Grids <: UltraDark.AbstractGrids\n\nGrids(length, resol::Int)\nGrids(length_tuple, resol_tuple::Tuple{Int, Int, Int})\n\nstruct containing grids used in a simulation\n\nFields\n\nx::Array{Float64, 3}: Array of x positions\ny::Array{Float64, 3}: Array of y positions\nz::Array{Float64, 3}: Array of z positions\nkx::Array{Float64, 3}: Array of x Fourier modes\nky::Array{Float64, 3}: Array of y Fourier modes\nkz::Array{Float64, 3}: Array of z Fourier modes\nk::Array{Float64, 3}: Fourier space postition array\nrkx::Array{Float64, 3}: Array of x Fourier modes for use with rfft\nrky::Array{Float64, 3}: Array of y Fourier modes for use with rfft\nrkz::Array{Float64, 3}: Array of z Fourier modes for use with rfft\nrk::Array{Float64, 3}: Fourier space postition array for use with rfft\nψx::Array{ComplexF64, 3}: ψ field\nψk::Array{ComplexF64, 3}: ψ field in Fourier space\nρx::Array{Float64, 3}: density field ρ\nρk::Array{ComplexF64, 3}: density field ρ in Fourier space\nΦx::Array{Float64, 3}: gravitational potential field Φ\nΦk::Array{ComplexF64, 3}: gravitational potential field Φ in fourier space\nfft_plan::Any: FFT plan for complex-to-complex transforms\nrfft_plan::Any: FFT plan for real-to-complex transforms\nk_vanish_indices::Vector{CartesianIndex{3}}: Indices at which k==0.0\nrk_vanish_indices::Any: Indices at which rk==0.0\n\nExamples\n\nCreate an empty grid with length length and resolution resol\n\njulia> using UltraDark\n\njulia> length = 1;\n\njulia> resol = 16;\n\njulia> Grids(length, resol);\n\njulia> size(ans.ψx)\n(16, 16, 16)\n\nCreate an empty length[1]xlength[2]xlength[3] grid with resolution resol[1]xresol[2]xresol[3].\n\njulia> using UltraDark\n\njulia> Grids((1.0, 1.0, 0.5), (64, 64, 32));\n\njulia> size(ans.ψx)\n(64, 64, 32)\n\n\n\n\n\n","category":"type"},{"location":"api/#UltraDark.PencilGrids","page":"API","title":"UltraDark.PencilGrids","text":"PencilGrids(length, resol)\nPencilGrids(length_tuple, resol_tuple::Tuple{Int, Int, Int})\n\nstruct containing grids used in a simulation\n\nEach grid is a PencilArray, allowing multiprocess FFTs. This comes with significant overhead so is only useful when running in a multi-node environment.\n\nFields\n\nx::Array{Float64, 3}: Array of x positions\ny::
[…]
