### 2024-04-15 b7d0e9b
doc: Move example outputs to tmpdir
Currently running the examples (and tests that include them) generates
data in the working directory.  Move this to temporary directories that
are cleaned up automatically on exit.
