### 2024-04-16 66e8aea
doc(joss): Fix typo in paper.md
