### 2024-04-16 e6e72f9
doc: Move `PencilGrids` example to Literate.jl
Literate.jl eases rendering scripts and their output in documentation
built by Documenter.jl.  This should make the examples much more
prominent and keep them up to date.

This commit refreshes the MPI/PencilGrids example and uses Literate.jl
to render it in the docs.
