+++ examples/soliton_velocity.jl
@@ -1,116 +1,139 @@
+# # PencilGrids example
+#
+# This example uses [`PencilGrids`](@ref UltraDark.PencilGrids) to demonstrate potential
+# use of [`MPI`](@extref MPI MPI.jl)-based parallelism.
+#
+# Note, however, that the online documentation is rendered by a single process and so using
+# `PencilGrids` has little practical benefit other than demonstrating how it is done.
+# If you run this example with multiple MPI processes, many of the `@info` statements will
+# print once for each process.
+#
+#
+# ## Initialise a simulation
+#
 using MPI
 using NPZ
 using CairoMakie
 using UltraDark
 
+# Initialise [`MPI`](@extref MPI MPI.jl) and set up [`MPI.Comm`](@extref) for inter-process communication.
 if ~MPI.Initialized()
     MPI.Init()
 end
 comm = MPI.COMM_WORLD
 @info "Process ID $(MPI.Comm_rank(comm)+1) of $(MPI.Comm_size(comm)) with $(Threads.nthreads()) threads\n"
 
+# Define grids on which simulation will be performed.
+# We use a relatively low resolution since this is just an example.
 const resol = 64
+const box_length = 10.0
 
-const num_snapshots = 20
-const output_times = LinRange(0, 5, num_snapshots)
+grids = PencilGrids(box_length, resol)
+@info "Grids created" MPI.Comm_rank(comm) size(grids.ψk)
 
+# Define initial conditions consisting of a soliton of mass `mass` at position `position0`
+# with velocity `velocity`.
 const mass = 10
 const position0 = [-4, 0, 0]
 const velocity = [1, 0, 0]
 const phase = 0
 const t0 = 0
 
-const output_dir = joinpath(mktempdir(), "output/vel/centered_on_soliton")
+UltraDark.Initialise.add_fdm_soliton!(grids, mass, position0, velocity, phase, t0)
+UltraDark.update_gravitational_potential!(grids) # ensure the density is up to date
+@info "Initialized" MPI.Comm_rank(comm) UltraDark.mass(grids)
 
-function run_sim()
+# Define output configuration
+const output_dir = joinpath(mktempdir(), "output")
 
-    output_config = OutputConfig(output_dir, output_times; box = true, slice = false)
-    options = Config.SimulationConfig()
+const num_snapshots = 20
+const output_times = LinRange(0, 5, num_snapshots)
 
-    grids = PencilGrids(10.0, resol)
-    @info "Grids created" MPI.Comm_rank(comm) size(grids.ψk)
+output_config = OutputConfig(output_dir, output_times; box = true, slice = false);
 
-    mkpath(output_config.directory)
-    npzwrite(joinpath(output_config.directory, "x.npy"), grids.x)
-    npzwrite(joinpath(output_config.directory, "y.npy"), grids.y)
-    npzwrite(joinpath(output_config.directory, "z.npy"), grids.z)
+# ## Run the simulation
+simulate!(grids, output_config) == nothing
+@info "Simulation complete" MPI.Comm_rank(comm)
 
-    UltraDark.Initialise.add_fdm_soliton!(grids, mass, position0, velocity, phase, t0)
-    @info "Initialized" MPI.Comm_rank(comm)
-
-    simulate!(grids, options, output_config) == nothing
-    @info "Simulation complete" MPI.Comm_rank(comm)
-
-end
-
-function plot_results()
-    @info "Starting plots" MPI.Comm_rank(comm)
-    pos_the =
-        reshape(position0, (1, :)) .+
[…]
