### 2024-04-16 #6035 @xuanxu
You can merge the JOSS branch into main before making the release, that way the new release based in main will include all the changes made during the review.
