### 2024-04-16 #6035 @musoke
@xuanxu, I have completed my checklist.  The details to report here are bolded:

    I have released **v0.9.7**
    - Julia General registry release: https://github.com/JuliaRegistries/General/commit/1701c161f8f2fa52fc82db68554120d7e899b7db, https://juliahub.com/ui/Packages/General/UltraDark/0.9.7
    - GitHub release: https://github.com/musoke/UltraDark.jl/releases/tag/v0.9.7
    I have archived the release on Zenodo with doi **https://doi.org/10.5281/zenodo.10978864**

The branch [`joss`](https://github.com/musoke/UltraDark.jl/tree/joss) is currently identical to tag [v0.9.7](https://github.com/musoke/UltraDark.jl/releases/tag/v0.9.7), the version archived at zenodo.  I will keep that branch around until this is done.
