### 2024-04-17 #6035 @musoke
The proof looks good to me!
